:sparkles: Add Frame logic
Frame logic encapsulates a pair of Rolls. Even though on a practical side the last frame of a Bowling games contains 3 Rolls, on the coding side it made more sense that the last Bowling game frame be represented by three frames max in case of three strikes, 2 frames in case of one strike/spare and one frame in case of no strike/spare.

Frame class also handles it's score calculation to ensure SRP.